small debug fixes

+++ source/memorySafeArray.cpp
@@ -174,8 +174,8 @@ Errors resizeSafeArray(SafeArray* array, size_t newSize) {
     // ASK: maybe it's better to clear even if we are resizing to a smaller size
 
     // KOLYA: ssize_t
-    size_t deltaSize = (newSize < oldSize ? oldSize - newSize : newSize - oldSize); // still cringe?
-    size_t deltaBytes = (size_t)deltaSize * array->elementSize;
+    size_t deltaSize  = (newSize < oldSize ? oldSize - newSize : newSize - oldSize); // still cringe?
+    size_t deltaBytes = deltaSize * array->elementSize;
     memset(array->array + oldSize + SIZE_OF_CANARY, 0, deltaBytes);
 
     Errors error = STATUS_OK;
